refactor(eval): share runtime option sanitizer
